requirements.txt
matplotlib==3.7.5
numpy==1.24.3
opencv-contrib-python==4.6.0.66
opencv-python==4.8.0.76
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.0.3
Pillow==9.5.0
pydantic==2.8.2